# SourceCode/swagger.yml
...
openapi: 3.0.0
info:
  title: "Trail microservice API"
  description: "An API for trails"
  version: "1.0.0"
servers:
  - url: "/api"
components:
  schemas:
    User:
      type: "object"
      properties:
        UserId:
          type: "integer"
        Email:
          type: "string"
        Role:
          type: "string"
...
  parameters:
...
    userId:
      name: "userId"
      description: "Id of the user"
      in: path
      required: True
      schema:
        type: "integer"
    email:
      name: "email"
      description: "email of the user"
      in: path
      required: True
      schema:
        type: "string"
paths:
...
  /users:
    get:
      operationId: "users.getAllUsers"
      tags:
        - Users
      summary: "Read the table of users"
      responses:
        "200":
          description: "Successfully read users table"
        "500":
          description: "Server error"

    post:
      operationId: "users.createNewUser"
      tags:
        - Users
      summary: "Create a user"
      requestBody:
        description: "User to create"
        required: True
        content:
           application/json:
            schema:
              x-body-name: "userJSON"
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: "Successfully created user"
        "400":
          description: "Bad request"
        "500":
          description: "Server error"
        
    put:
      tags:
        - Users
      operationId: "users.updateUserById"
      summary: "Update a user"
      requestBody:
        description: "User to update"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "userJSON"
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: "Successfully updated user"
        "400":
          description: "Bad request"
        "404":
          description: "User not found"
  /users/{userId}:

    get:
      operationId: "users.getUserById"
      tags:
        - Users
      summary: "Read specified user by id"
      parameters:
       - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: "Successfully read specified user"
        "400":
          description: "Bad request"
        "404":
          description: "User not found"
        "500":
          description: "Server error"

    delete:
      operationId: "users.deleteUserById"
      tags:
        - Users
      summary: "Delete a user with given ID"
      parameters:
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: "Successfully deleted user"
        "400":
          description: "Bad request"
        "404":
          description: "User not found"
        "500":
          description: "Server error"
  /users/email/{email}:
    get:
      operationId: "users.getIdByEmail"
      tags:
        - Users
      summary: "Get specified id from email"
      parameters:
       - $ref: "#/components/parameters/email"
      responses:
        "200":
          description: "Successfully found user"
        "400":
          description: "Bad request"
        "404":
          description: "User not found"
        "500":
          description: "Server error"
...